Replace '!' with "!".
The char constant '!' was replaced by "!" because the call of the operator *+=*
applied on a UTF8string object and a constant char does not make sense.

A character is a byte from 0 to 256. So, adding a char *c > 127* in a UTF-8
encoded string results automatically in the *std::invalid_argument* exception.
So I decided to remove this operator (next commit).