utf8_iterator : Fixed cppcheck issues.
Firstly, in the definition of the overloaded operator '=', the index of
*this* was not assigned to the index of the iterator given as argument.

Secondly, the constructor was not declared as an *explicit* constructor.
That is to say it was possible to write this kind of code:

	UTF8iterator(std::string("a utf8 string"));

or (it may be possible):

    UTF8iterator("a utf8 string"); (2)

instead of:

    UTF8iterator(UTFstring("a utf8 string"));

(1) is possible because there is a constructor of *UTF8string* with
*std::string* as argument. So an implicit conversion is possible.

(2) is also possible for similiar reasons.
But it is possible the compiler refuses that piece of code and
returns an error. Some compilers can accept that and compute an
implicit conversion from *std::string* to *UTF8iterator*. It is
possible by transitivity.
*std::string* as a constructor with a pointer to char as argument.
And lkie I said in (1), *UTF8string* has a constructor
with *std::string* as argument. So by transitivity:

    std::string -> UTF8string   ≡ ①
	UTF8string  -> UTF8iterator ≡ ②

	Hence, ① ∧ ② ≡ std::string -> UTF8iterator
	(Α -> Β) means "Β can be build from Α"

Consequently, with some compilers, an implicit conversion can be done
from std::string to UTF8iterator in some programs that use it.
So, that should be a bug.
That is why cppcheck noticed that.

Author: Luxon JEAN-PIERRE <luxon.jean.pierre@gmail.com>