utf8_string: basic_string<unsigned char>-based UTF8string
I rebuilt the internal part of UTF8string because my internal string
(in UTF-8) is stored in an ```std::string``` which is an alias of
```basic_string<char>```.
It was a problem because of the implicit conversions
used by the compiler when I compare my codepoints of a multi-byte character
to validate a utf-8 string.

Fix issue #2

--